image_encrypt/image_encrypt/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
cloudinary==1.41.0
python-multipart==0.0.12
